denim-agent-backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlmodel>=0.0.16
pydantic>=2.6.4
python-dotenv>=1.0.1
exa-py>=1.0.7
httpx>=0.27.0
beautifulsoup4>=4.12.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
arq>=0.25.0
redis>=5.0.0
pydantic-settings>=2.0.0
bcrypt<4.0.0